add move constructor in class Message and move the operation credential from the ipc service to the worker

+++ src/client_lib/src/ipc/src/lib/callback/RequestContext.h
@@ -19,7 +19,7 @@ public:
 class ReadRequestContext : public RequestContext{
 
 public:
-	uint32_t remainSize_;
+	uint64_t remainSize_;
 
 	IPCWriteRequestMessage lastResponse_;
